+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - close active editor on create
 - [strip managedFields](https://github.com/kubernetes/kubernetes/pull/96878)
+
+### Added
 - PVC -> right click -> Go to PV
+- reveal
 
 ## [0.1.1] - 2021-05-13
 ### Changed