+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PVC -> right click -> Go to PV
 - reveal
 - namespace -> right click -> edit
+- right click on pod -> kubectl exec, kubectl debug
 ### Fixed
 - metadata.selfLink is deprecated